Topic/shldr mnt arms (#1)
Feature: Set the gripper constraints (dynamics) by loading geometry type from json file.

The default geometry type is cylinder, but setting the geometry_type in the json file to box will load a box style geometry and will set the size using the size field json file.

The gripper constraints are modified to grasp the box-style geometry if the object is of ST.box type. Otherwise, the gripper is set for the cylinder geometry as default.

The manipulator configurations are loaded from the json configs meaing that they can be moved via json rather than from the .g file as was prior to this feature.

This feature is tested using the environments: square and extreme_beam_test. The former is a new configuration for dual-arm assembly, whilst the latter is previous configuration. Both work as expected.